Reorg repo for v1 and v2 approaches
