+++ proteogram/v1/__init__.py
@@ -1,0 +1,4 @@
+from .image_similarity import Img2Vec
+from .proteogram import Proteogram
+
+__all__ = ['Proteogram', 'Img2Vec']
