+++ scripts/utilities/copy_structures.py
