+++ scripts/utilities/copy_structures_by_prefix.py
