+++ scripts/utilities/find_structures_in_scope.py
