+++ scripts/v1/make_training_data_exclude_eval.py
@@ -16,8 +16,8 @@ import warnings
 from Bio.SCOP import Scop
 from Bio.PDB.PDBParser import PDBConstructionWarning
 
-from proteogram.utils import read_yaml
-from proteogram.proteogram import Proteogram
+from proteogram.common.utils import read_yaml
+from proteogram.v1.proteogram import Proteogram
 
 
 # Distance cutoff for measuring possible residue interactions in Angstroms
